Checking in, but Legend is not completely done yet: Font sizes are adjusted propperly, but legend widget size is not adjusted to changing font sizes yet.
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@722 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ src/KDChartAbstractCartesianDiagram.cpp
@@ -31,9 +31,9 @@
 
 #include <cmath>
 
-static int _trunc( qreal v )
+static qreal _trunc( qreal v )
 {
-    return static_cast<int>( ( v > 0 ) ? floor( v ) : ceil(  v ) );
+    return (( v > 0.0 ) ? floor( v ) : ceil(  v ));
 }
 
 using namespace KDChart;
